+++ src/main.hs
@@ -1,15 +1,14 @@
-{-# LANGUAGE ApplicativeDo #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards #-}
+module Main where
 
-import Common.Options (Options (..))
-import Data.Foldable (fold)
+import Common.Options (subparsers)
+import Control.Monad (join)
 import Data.Version (showVersion)
 import Options.Applicative
 import PackageInfo_cardano_ledger_release_tool (version)
 
 import qualified Changelogs.Command as Changelogs
 import qualified System.Console.Terminal.Size as TS
+import qualified Workflow.Command as Workflow
 
 main :: IO ()
 main = do
@@ -23,23 +22,13 @@ main = do
           <> long "version"
           <> hidden
 
-  (subcmd, opts) <-
-    customExecParser
-      (prefs $ columns cols)
-      ( info
-          ( helper <*> versionOption version <*> do
-              optVerbose <-
-                switch $
-                  help "Produce verbose output"
-                    <> short 'v'
-                    <> long "verbose"
-              subcmd <-
-                subparser . fold $
-                  [ Changelogs.subcmd
-                  ]
-              pure (subcmd, Options {..})
-          )
-          (fullDesc <> header "Cardano Ledger release tool")
+  join . customExecParser (prefs $ columns cols) $
+    info
+      ( helper
+          <*> versionOption version
+          <*> subparsers
+            [ Changelogs.subcmd
+            , Workflow.subcmd
+            ]
       )
-
-  subcmd opts
+      (fullDesc <> header "Cardano Ledger release tool")