+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version history for `cardano-ledger-release-tool`
 
 ## 0.2.0.0
 
+* Rename the executable to `cleret`
 * Add `-V` as a short-option form of `--version`
 
 ## 0.1.1.0