+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version history for `cardano-ledger-release-tool`
 
 ## 0.2.0.0
 
+* Add a `cabal` group of subcommands
 * Add a `workflow` group of subcommands
 * Make the `changelogs` subcommand a group
 * Rename the executable to `cleret`